refactor: ReservationDetail -> MovieDetail 클래스 명 변경, finished 패키지 삭제
