+++ README.md
@@ -11,7 +11,7 @@
   - [x] 포스터  
   - [x] 지금 예매 버튼
 
-### 영화 예매
+### 영화 상세
 - [x] 툴바
   - [x] 제목
   - [x] 뒤로 가기 버튼
@@ -53,7 +53,7 @@
 - [x] 영화 목록에서 지금 예매 버튼을 누르면 영화 예매 화면으로 이동
   - [x] 이동 시 선택된 아이템에 해당하는 영화 정보 전달
 
-### 영화 예매 화면
+### 영화 상세 화면
 - [ ] 영화 예매에서 좌석 선택 버튼을 누르면 좌석 선택 화면으로 이동
   - [x] 티켓 매수 전달 
   - [x] 영화 정보 전달
