+++ app/src/main/AndroidManifest.xml
@@ -23,12 +23,12 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".reservation.finished.ReservationFinishedActivity"
+            android:name=".reservation.ReservationFinishedActivity"
             android:parentActivityName=".home.HomeActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".reservation.detail.ReservationDetailActivity"
+            android:name=".detail.MovieDetailActivity"
             android:parentActivityName=".home.HomeActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
