+++ app/src/main/java/woowacourse/movie/reservation/ReservationFinishedPresenter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.reservation.finished
+package woowacourse.movie.reservation
 
 import woowacourse.movie.db.Movies
 import woowacourse.movie.model.Ticket
