+++ app/src/main/res/layout/activity_reservation_finished.xml
@@ -5,7 +5,7 @@
     android:id="@+id/constraint_layout_reservation_finished"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".reservation.finished.ReservationFinishedActivity">
+    tools:context=".reservation.ReservationFinishedActivity">
 
     <View
         android:id="@+id/view_reservation_finished"
